break down project test spec into multiple files
