+++ ai-verify-portal/__tests__/modules/project.canvas.test.tsx
@@ -416,90 +416,4 @@ describe('Project Flow', () => {
       expect(gridLayoutArea.children.length).toBe(0); //remaining page has no widget
     });
   });
-
-  describe('Datasets and Model Selection Screen', () => {
-    it('should render Datasets / Model selection screen', async () => {
-      const fetchSpy = jest.spyOn(global, 'fetch');
-      fetchSpy.mockImplementation(
-        jest.fn(() =>
-          Promise.resolve({
-            json: () => Promise.resolve(fmtInterpretationMAEBundleResponse),
-          })
-        ) as jest.Mock
-      );
-      const user = userEvent.setup();
-      const { container } = render(
-        <MockProviders apolloMocks={mockGqlDataMinimal}>
-          <ProjectCreatePageWrapper />
-        </MockProviders>
-      );
-      const projNameInput: HTMLInputElement =
-        await screen.findByPlaceholderText(
-          /^Enter name of this project e.g. Credit Scoring Model Tests$/i
-        );
-      await userEvent.type(projNameInput, 'Test Project');
-      const nextBtn = await screen.findByText(/^Next$/i);
-      await user.click(nextBtn.parentElement as HTMLDivElement);
-      await screen.findByText(/^Select Report Template$/i);
-      const blankCard = await screen.findByText(/^Blank Canvas$/i);
-      await userEvent.click(blankCard);
-      await waitFor(() => {
-        expect(container.querySelector('.card__highlighted')).toBeDefined();
-      });
-      await userEvent.click(nextBtn);
-      await screen.queryByText(/^Design Report$/i);
-      await screen.findByText(/^Interpretation \(MAE\)$/i);
-      const gridLayoutArea = container.querySelector(
-        GRID_LAYOUT_CLASSNAME
-      ) as HTMLDivElement;
-      expect(gridLayoutArea.children.length).toBe(0);
-      const maeInterWidget = container.querySelector(
-        '[data-test-id="draggableWidget-aiverify.stock.fairness_metrics_toolbox_for_regression:mae_interpretation"]'
-      );
-      fireEvent.dragStart(maeInterWidget as Element, {
-        dataTransfer: { setData: (_str: string) => undefined },
-        clientX: 518,
-        clientY: 326,
-        screenX: -1402,
-        screenY: 435,
-        x: 374,
-        y: 218,
-      });
-
-      fireEvent.dragOver(gridLayoutArea as Element, {
-        clientX: 518,
-        clientY: 326,
-        screenX: -1402,
-        screenY: 435,
-        x: 374,
-        y: 218,
-      });
-
-      await waitFor(() => {
-        expect(container.querySelector(GRID_ITEM_CLASSNAME)).not.toBeNull();
-      });
-
-      fireEvent.drop(gridLayoutArea as Element, {
-        clientX: 518,
-        clientY: 326,
-        screenX: -1402,
-        screenY: 435,
-        x: 374,
-        y: 218,
-      });
-
-      //widget should be on canvas
-      await waitFor(() => {
-        expect(container.querySelector('.widgetContainer')).not.toBeNull();
-      });
-
-      await user.click(await screen.findByText(/^Next$/i));
-      await screen.findByText(
-        /^Select the Datasets and AI Model to be tested$/i
-      );
-      expect(container.querySelector('.layoutContentArea')).toMatchSnapshot(
-        'Datasets/Model-Selection-Screen'
-      ); // snapshot without header, because header has dynamic autosave time display
-    });
-  });
 });
